+++ src/lib.rs
@@ -40,12 +40,3 @@ pub use crate::bytes::WeakBytes;
 pub use crate::pyanybytes::PyAnyBytes;
 #[cfg(feature = "zerocopy")]
 pub use crate::view::View;
-
-/// Erase the lifetime of a reference.
-///
-/// # Safety
-/// The caller must guarantee that the referenced data remains valid for the
-/// `'static` lifetime.
-unsafe fn erase_lifetime<'a, T: ?Sized>(slice: &'a T) -> &'static T {
-    &*(slice as *const T)
-}